feat: added db context to use sqlite - bstone
