+++ src/Controllers/UserController.cs
@@ -1,0 +1,125 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
+using NoteAppApi.Models;
+
+//aspnet-codegenerator controller -name UsersController -async -api -m User -dc NoteContext -outDir Controllers
+
+namespace NoteAppApi.Controllers
+{
+	[Route("api/[controller]")]
+	[ApiController]
+	public class UserController : ControllerBase
+	{
+		private readonly NoteContext _context;
+
+		public UserController(NoteContext context)
+		{
+			_context = context;
+		}
+
+		// GET: api/User
+		[HttpGet]
+		public async Task<ActionResult<IEnumerable<User>>> GetUsers()
+		{
+			if (_context.Users == null)
+			{
+				return NotFound();
+			}
+			return await _context.Users.ToListAsync();
+		}
+
+		// GET: api/User/5
+		[HttpGet("{id}")]
+		public async Task<ActionResult<User>> GetUser(int id)
+		{
+			if (_context.Users == null)
+			{
+				return NotFound();
+			}
+			var user = await _context.Users.FindAsync(id);
+
+			if (user == null)
+			{
+				return NotFound();
+			}
+
+			return user;
+		}
+
+		// PUT: api/User/5
+		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		[HttpPut("{id}")]
+		public async Task<IActionResult> PutUser(int id, User user)
+		{
+			if (id != user.Id)
+			{
+				return BadRequest();
+			}
+
+			_context.Entry(user).State = EntityState.Modified;
+
+			try
+			{
+				await _context.SaveChangesAsync();
+			}
+			catch (DbUpdateConcurrencyException)
+			{
+				if (!UserExists(id))
+				{
+					return NotFound();
+				}
+				else
+				{
+					throw;
+				}
+			}
+
+			return NoContent();
+		}
+
+		// POST: api/User
+		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		[HttpPost]
+		public async Task<ActionResult<User>> PostUser(User user)
+		{
+			if (_context.Users == null)
+			{
+				return Problem("Entity set 'NoteContext.Users'  is null.");
+			}
+			_context.Users.Add(user);
+			await _context.SaveChangesAsync();
+
+			return CreatedAtAction("GetUser", new { id = user.Id }, user);
+		}
+
+		// DELETE: api/User/5
+		[HttpDelete("{id}")]
+		public async Task<IActionResult> DeleteUser(int id)
+		{
+			if (_context.Users == null)
+			{
+				return NotFound();
+			}
+			var user = await _context.Users.FindAsync(id);
+			if (user == null)
+			{
+				return NotFound();
+			}
+
+			_context.Users.Remove(user);
+			await _context.SaveChangesAsync();
+
+			return NoContent();
+		}
+
+		private bool UserExists(int id)
+		{
+			return (_context.Users?.Any(e => e.Id == id)).GetValueOrDefault();
+		}
+	}
+}
